Fix for possible race condition when writing token in the disk cache
The disk cache is used to cache the token. The problem arises when multiple requests are handled concurrently. Imagine, there is no token in the cache at the moment. Hence `get` in disk cache returns `null` for the first request causing it to download a fresh token. While the file is written, another request reads the file at the very same moment, resulting in undefined and possible invalid token values and sporadic (inexplicable) error messages.

This commit wraps the critical section into a mutex and grants exclusive access to the file, in case, it is written.